[ANDROSDK-1419] Ktlint
