+++ core/src/main/java/org/hisp/dhis/android/core/analytics/LegendEvaluator.kt
@@ -81,6 +81,7 @@ internal class LegendEvaluator @Inject constructor(
         }
     }
 
+    @Suppress("UnusedPrivateMember", "FunctionOnlyReturningConstant")
     fun getLegendByTrackedEntityAttribute(
         trackedEntityAttributeUid: String,
         value: String?
